fix: real build failures in journey-service, transaction-service, iot-service
The first CI run after adding these services to the build matrix (run 33771816075)
surfaced that most of them had never actually been built before. Root-caused and
fixed each one by reproducing the exact CI steps locally via nerdctl:

journey-service, transaction-service (Dockerfiles):
  Both used a broken two-stage `mvn ... go-offline -o` pattern -- `-o` (offline
  mode) on the go-offline step itself, before anything had been fetched into a
  cold ~/.m2 cache, so it could never resolve even the parent POM. Simplified
  to the same single-step `mvn package` pattern identity-service/workflow-service
  already use successfully.

transaction-service (real compile errors, unrelated to Docker):
  - OrderItem.java used UUID.randomUUID() with no `import java.util.UUID`.
  - CartService.listCarts() called CartRepository.findByIdentityIdOrderByCreatedAtDesc,
    which didn't exist on the repository -- added it.
  - FulfillmentScheduler called OrderRepository's 2-arg method signature, but the
    repository only declares a 1-arg version (status filter is already hardcoded
    in its @Query) -- fixed the call site to match.

iot-service (go.mod/go.sum never existed, and nothing had ever compiled):
  - go.mod was missing github.com/lib/pq and go.opentelemetry.io/collector/pdata
    entirely despite both being imported, referenced confluent-kafka-go/v2 at a
    version (v2.5.1) that was never tagged, and had no go.sum committed at all.
    Regenerated via `go mod tidy` against a real toolchain -- pdata v1.66.0 pulls
    in a `go 1.26` requirement, so bumped the go directive and go.sum accordingly.
  - internal/fleet/manager.go: `rand.Intn(1000)` assigned into a `map[string]float64`
    with no float64(...) conversion.
  - internal/kafka/producer.go: called Producer.Produce with a callback function
    as the second argument; the real API takes `chan Event`. Switched both call
    sites to a delivery channel + goroutine that logs the same way the code
    originally intended.
  - internal/ingestion/otlp.go: `range` over a pmetric.ResourceMetricsSlice (not
    rangeable -- needs Len()/At(i)), `Metric.MetricType()` doesn't exist (real
    method is `Type()`), plus two unused imports.
  - internal/anomaly/detector.go: unused loop variable.
  - cmd/iot-service/main.go: called anomaly.NewDetector with 1 arg instead of the
    real 2-arg signature, and referenced an undefined `buildRouter` function --
    the real router (internal/api.NewRouter(...).Handler()) was already
    implemented but never wired in.
  - Dockerfile: confluent-kafka-go's bundled librdkafka static archive is built
    against glibc (undefined setcontext/getcontext/fopen64 when linked with
    musl's ld) -- switched to the `dynamic` build tag + apk librdkafka-dev/
    pkgconfig, which links against a real system librdkafka instead. Also bumped
    the runtime stage from alpine:3.20 to 3.24 to match golang:1.26-alpine's
    Alpine release: 3.20's librdkafka (2.4.0) is older than what the client
    requires (>= 2.5.3), so the binary refused to start with a version-mismatch
    error even after it linked successfully.

Also added the missing package-lock.json for weather-service and news-ticker-service
(neither had one committed, so CI's `npm ci` step -- which requires an existing
lockfile -- failed before ever reaching the Docker build) and bumped build.yml's
shared Go setup version from 1.22 to 1.26 for the whole build-go matrix (safe: the
other 3 Go services only declare `go 1.22`, which builds fine under a newer
toolchain).

Verified end-to-end locally via nerdctl (mirroring the actual CI steps): all three
services now build and run -- iot-service was actually started in a container and
confirmed to bring up the Kafka producer, OTLP gRPC/HTTP receivers, REST API, and
fleet simulator cleanly (only expected connection-refused noise against fake
Kafka/Postgres endpoints in the test environment).

+++ services/iot-service/internal/fleet/manager.go
@@ -104,7 +104,7 @@ func (fm *FleetManager) updateMetrics(device *Device) {
 		device.Metrics["co2_ppm"] = 400 + rand.Float64()*800 // 400-1200 ppm
 	case DeviceTypeMachine:
 		device.Metrics["vibration"] = 0.1 + rand.Float64()*1.5 // 0.1-1.6 mm/s
-		device.Metrics["cycle_count"] = rand.Intn(1000)
+		device.Metrics["cycle_count"] = float64(rand.Intn(1000))
 		device.Metrics["temp"] = 20 + rand.Float64()*80 // 20-100°C
 		device.Metrics["error_rate"] = rand.Float64() * 10 // 0-10%
 		device.Metrics["throughput"] = 50 + rand.Float64()*150 // 50-200 units/min
